% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aggregate_data.R
\name{fct_aggregate_data}
\alias{fct_aggregate_data}
\title{fct_aggregate_data}
\usage{
fct_aggregate_data(
  df5W,
  lookup_dfindicator,
  proportions = "pin",
  totalmodel = "sum"
)
}
\arguments{
\item{df5W}{fame with response tracking data}

\item{lookup_dfindicator}{fame with indicator look up}

\item{proportions}{can be "pin" or "target" - this uses configuration available 
within the spreadsheet inst/Proportions.xlsx}

\item{totalmodel}{can be either: 
*  "sum":  sums all beneficiaries of all sectors to get
 intersector figures per admin1 level and then at national level

*  "maxsector" : sums all beneficiaries at sector and admin1
  level, then take the max of each age and gender categories
   for each population type to get the intersector figure 
 
*  "southernconemodel" : Mixed approach in 3 steps: 

1. Max across Shelter, Food security, Humanitarian transport
    and WASH 
    
2. Take Protection (General) data  

3.  Sum Other sectors mentionned: Integration, 
    Multipurporse CBI, Health, Education}
}
\value{
list with results
}
\description{
Function to aggregate data according to different scenario
}
\examples{

lookup_dfindicator <- fct_read_lookup(type = "indicator")
df5W <- fct_read_data() 


df5Wctr <- fct_read_data(filter = "country",
                         value = "Bolivia")
resultpinsum <- fct_aggregate_data(df5W = df5Wctr,
                             lookup_dfindicator, 
                             proportions = "pin",  
                             totalmodel = "sum" )
## Check all 6 combinations...

### Combination 1
resultpinsum <- fct_aggregate_data(df5W,
                             lookup_dfindicator, 
                             proportions = "pin",  
                             totalmodel = "sum"   
)

head(resultpinsum, 10)

### Combination 2
resulttargetsum <- fct_aggregate_data(df5W,
                             lookup_dfindicator,
                             proportions = "target",  
                             totalmodel = "sum"   
)

head(resulttargetsum, 10)


### Combination 3
resultpinmaxsector <- fct_aggregate_data(df5W,
                             lookup_dfindicator,
                             proportions = "pin",  
                             totalmodel = "maxsector"   
)

head(resultpinmaxsector, 10)


### Combination 4
resulttargetmaxsector<- fct_aggregate_data(df5W,
                             lookup_dfindicator,
                             proportions = "target",  
                             totalmodel = "maxsector"   
)

head(resulttargetmaxsector, 10)


### Combination 5
resultpinsouthernconemodel <- fct_aggregate_data(df5W,
                             lookup_dfindicator,
                             proportions = "pin",  
                             totalmodel = "southernconemodel"   
)

head(resultpinsouthernconemodel, 10)


### Combination 6
resulttargetsouthernconemodel <- fct_aggregate_data(df5W,
                             lookup_dfindicator,
                             proportions = "target", 
                             totalmodel = "southernconemodel"   
)

head(resulttargetsouthernconemodel, 10)




}
