% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_error_report.R
\name{fct_error_report}
\alias{fct_error_report}
\title{fct_error_report}
\usage{
fct_error_report(result, countryname = NULL)
}
\arguments{
\item{result}{output of the previous function fct_read_data()}

\item{countryname}{name of the country to filter the report
can be set to "All" - or any of  "Aruba", "Curaçao",
 "Costa Rica", "Dominican Republic",
 "Trinidad and Tobago", "Guyana", 
              "Mexico", "Panama"}
}
\value{
result list with results and data error...
}
\description{
The functions perform systematic Data Quality Check
}
\details{
Currently this includes the 19 following Check: 

 __Completeness__:

 1.  missingcountry: Country information missing
 2. countryadmincheck: Administrative unit level1 not matching 
 3.  admin1and2check: Administrative unit level2 not matching 
 4.  miss_appeal_org: Missing appealing organsation
 5.  miss_setup:   Missing Implementation
 6.  miss_implementing_org:   Missing name of implementer
 7.  miss_month: Missing month
 8.  missing_what,
  
 __Consistency__:
 9.  wrongsectindicator, 
 10.  zeroCVA, 
 11.  missingmechanism, 
 12.  CVANotoYes,         
 13.  MultipurposeSector, 
 14.  DirectAssistanceNoBenef, 
 15.  NewBenefvstotal, 
 
 __Accuracy__:
 16.  PopTypeBreakdown, 
 17.  AGDBreakdown, 
 18.  CBuildingNoBenef, 
 19.  NoOutput
}
\examples{
## Get data
result0 <- fct_read_data()
## Fix errors
result <- fct_error_report(result = result0,
                             countryname = NULL)
 
## Use parameters
result <- fct_error_report(result = result0,
                           countryname ="All")
## charts 
print(result$plot_Country)
result <- fct_error_report(result = result0,
                           countryname ="Colombia")

print(result$plot_Appealing)
## Error
head(result[["Error"]], 10)
## Display fixed data 
head(result[["ErrorReportclean"]], 10)


}
