% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{fct_read_data}
\alias{fct_read_data}
\title{fct_read_data}
\usage{
fct_read_data(filter = NULL, value = NULL)
}
\arguments{
\item{filter}{what to filter on (country or partner) so that we do not need to load everything}

\item{value}{value to filter from}
}
\value{
df5W  response tracking potentially filtered
}
\description{
Pull Data From ActivityInfo
}
\details{
Note that this function is specific to a specific DB schema -
the form to query as well as the columns variabl name are hard coded and 
would need to be adjusted if the DB change

Note also that the function is based on a token stored as an environment 
variable

 # Credentials set up as environment variables - 
 
 # token <- "activityinfotoken.."
 # print(Sys.setenv(ACTIVITYINFOTOKEN = token))   
 # Sys.getenv("ACTIVITYINFOTOKEN")
 # rm(token)
}
\examples{

## Pulling everything... can be long..
df5W <- fct_read_data()
head( df5W, 10)

## testing the filters
df5Wctr <- fct_read_data(filter = "country",
                         value = "Colombia")
head( df5Wctr, 10)

df5Wpart <- fct_read_data(filter = "partner",
                         value = "United Nations High Commissioner for Refugees (UNHCR)")
head( df5Wpart , 10)

}
