% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aggregate_data.R
\name{fct_aggregate_data}
\alias{fct_aggregate_data}
\title{fct_aggregate_data}
\usage{
fct_aggregate_data(
  result,
  countryname = NULL,
  proportions = "pin",
  totalmodel = "sum"
)
}
\arguments{
\item{result}{output of the previous function fct_read_data() 
and fct_error_report()}

\item{countryname}{name of the country to filter the report
can be set to "All" - or any of  "Aruba", "Curaçao",
 "Costa Rica", "Dominican Republic",
 "Trinidad and Tobago", "Guyana", 
              "Mexico", "Panama"}

\item{proportions}{can be "pin" or "target" - this uses configuration available 
within the spreadsheet inst/Proportions.xlsx}

\item{totalmodel}{can be either: 
*  "sum":  sums all beneficiaries of all sectors to get
 intersector figures per admin1 level and then at national level

*  "maxsector" : sums all beneficiaries at sector and admin1
  level, then takethe max of each age and gender categories
   for each population type to get the intersector figure 
 
*  "southernconemodel" : Mixed approach in 3 steps: 

1.Max across Shelter, Food security, Humanitarian transport
    and WASH 
    
2. Take Protection (General) data  

3.  Sum Other sectors mentionned: Integration, 
    Multipurporse CBI, Health, Education}
}
\value{
list with results
}
\description{
Function to aggregate data according to different scenario
}
\examples{

result <- fct_read_data()
## Fix errors
result2 <- fct_error_report(result)

## Check all 6 combinations...

### Combination 1
resultpinsum <- fct_aggregate_data(result = result2,
                             countryname = NULL, 
                             proportions = "pin",  
                             totalmodel = "sum"   
)

head(resultpinsum[["ConsolidatedReport"]], 10)

### Combination 2
resulttargetsum <- fct_aggregate_data(result2,
                             countryname = NULL, 
                             proportions = "target",  
                             totalmodel = "sum"   
)

head(resulttargetsum[["ConsolidatedReport"]], 10)


### Combination 3
resultpinmaxsector <- fct_aggregate_data(result2,
                             countryname = NULL, 
                             proportions = "pin",  
                             totalmodel = "maxsector"   
)

head(resultpinmaxsector[["ConsolidatedReport"]], 10)


### Combination 4
resulttargetmaxsector<- fct_aggregate_data(result2,
                             countryname = NULL, 
                             proportions = "target",  
                             totalmodel = "maxsector"   
)

head(resulttargetmaxsector[["ConsolidatedReport"]], 10)


### Combination 5
resultpinsouthernconemodel <- fct_aggregate_data(result2,
                             countryname = NULL, 
                             proportions = "pin",  
                             totalmodel = "southernconemodel"   
)

head(resultpinsouthernconemodel[["ConsolidatedReport"]], 10)


### Combination 6
resulttargetsouthernconemodel <- fct_aggregate_data(result2,
                             countryname = NULL, 
                             proportions = "target", 
                             totalmodel = "southernconemodel"   
)

head(resulttargetsouthernconemodel[["ConsolidatedReport"]], 10)




}
