% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_error_report.R
\name{fct_error_report}
\alias{fct_error_report}
\title{fct_error_report}
\usage{
fct_error_report(result, countryname = NULL)
}
\arguments{
\item{result}{output of the previous function fct_read_data()}

\item{countryname}{name of the country to filter the report
can be set to "All" - or any of  "Aruba", "Curaçao",
 "Costa Rica", "Dominican Republic",
 "Trinidad and Tobago", "Guyana", 
              "Mexico", "Panama"}
}
\value{
result list with results and data error...
}
\description{
The functions perform systematic Data Quality Check
}
\examples{
## Get data
result <- fct_read_data()
## Fix errors
result <- fct_error_report(result)
## Display fixed data 

head(result[["ErrorReportclean"]], 10)

}
