% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{fct_read_data}
\alias{fct_read_data}
\title{fct_read_data}
\usage{
fct_read_data()
}
\value{
list with the content of the activity info DB 
               * df5W 
               * dfadmin1  
               * dfadmin2  
               * dfindicator   
               * dfpartner
}
\description{
Pull Data From ActivityInfo
}
\details{
Note that this function is specific to a specific DB schema -
the form to query as well as the columns variabl name are hard coded and 
would need to be adjusted if the DB change

Note also that the function is based on a token stored as an environment 
variable

 # Credentials set up as environment variables - 
 
 # token <- "activityinfotoken.."
 # print(Sys.setenv(ACTIVITYINFOTOKEN = token))   
 # Sys.getenv("ACTIVITYINFOTOKEN")
 # rm(token)
}
\examples{
result <- fct_read_data()
## list of results

head(result[["df5W"]], 10)
head(result[["dfadmin1"]], 10)
head(result[["dfadmin2"]], 10)
head(result[["dfindicator"]], 10)
head(result[["dfpartner"]], 10)
}
